update vfclass_v insn.

+++ riscv/insns/vfclass_v.h
@@ -1,5 +1,5 @@
 require_extension('V');
 check_vstart{
-	Vfclass<unsigned int, unsigned short> vfclass;
-	vfclass.vfclass_v((half*)&VRS2.vh[VSTART], (unsigned int*)&VRD.vh[VSTART], VM, VR0.vh, VL);
+	Vfclass<unsigned short, unsigned short> vfclass;
+	vfclass.vfclass_v((half*)&VRS2.vh[VSTART], (unsigned short*)&VRD.vh[VSTART], VM, VR0.vh, VL);
 }
